\name{all_besags_l}
\alias{all_besags_l}
\title{Make sure that only the non-negative functions will get Besag's
L-transform.}
\usage{
  all_besags_l(all_k_a, is_any_mark_neg, is_any_mark_pos)
}
\arguments{
  \item{all_k_a}{An array containing all K_f for all
  patterns. Dimensions: orig_and_perm, summ_func, r.}
}
\description{
  Make sure that only the non-negative functions will get
  Besag's L-transform.
}

