% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mtf.r
\name{besags_L_valid}
\alias{besags_L_valid}
\title{Is L-transform valid?}
\usage{
besags_L_valid(mtf_name, is_any_mark_neg, is_any_mark_pos)
}
\arguments{
\item{mtf_name}{A vector of mark test function names. "1" stands for the
unmarked K-function. Accepted values are '1', 'm', 'mm', 'gamma',
'gammaAbs' and 'morAbs'.}

\item{is_any_mark_neg}{A logical scalar value describing whether any of
the marks in the mark distribution have a negative value.}

\item{is_any_mark_pos}{A logical scalar value describing whether any of
the marks in the mark distribution have a positive value.}
}
\value{
A vector of the names of those mark test function names which
  have a valid Besag's L-transform. The names will be in the same order
  as they appear in mtf_name.
}
\description{
Ask for which mark test functions Besag's L-transformation can be done
given a mark distribution.
}
\details{
Given names of mark test functions and information on the (one-point)
mark distribution, returns the names of those mark test functions which
have a valid Besag's L-transform for the specified mark distribution.
}

