% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mtf.r
\name{check_mtf}
\alias{check_mtf}
\title{Check that the given mark test function names are valid.}
\usage{
check_mtf(mtf_name)
}
\arguments{
\item{mtf_name}{A string giving the name of the mark test function.
Possible names are '1', 'm', 'mm', 'gamma', 'gammaAbs', 'morAbs'.}
}
\description{
Check that the given mark test function names are valid.
}

