\name{check_or_calc_r_max}
\alias{check_or_calc_r_max}
\title{If r_max has not been given, recommend a maximum radius based on the
window.}
\usage{
  check_or_calc_r_max(r_max, min_window_side,
    recommended_r_max_ratio = 0.25)
}
\description{
  If r_max has not been given, recommend a maximum radius
  based on the window.
}

