% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/summ-func.r
\name{check_pattern}
\alias{check_pattern}
\title{Checks that the given pattern is valid.}
\usage{
check_pattern(pattern)
}
\arguments{
\item{pattern}{A pattern to be checked.}
}
\description{
Checks that the given pattern is valid.
}

