% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/radius.r
\name{check_r_max}
\alias{check_r_max}
\title{Check that the given maximum radius makes sense.}
\usage{
check_r_max(r_max, min_window_side)
}
\arguments{
\item{r_max}{Given value for the maximum radius.}

\item{min_window_side}{The minimum window side length.}
}
\description{
Check that the given maximum radius makes sense.
}

