% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radius.r
\name{check_r_vec}
\alias{check_r_vec}
\title{Check that the given radius vector makes sense.}
\usage{
check_r_vec(r_vec, min_window_side)
}
\arguments{
\item{r_vec}{Given radius vector.}

\item{min_window_side}{The minimum window side length.}
}
\description{
Check that the given radius vector makes sense.
}

