\name{create_mark_test_funcs}
\alias{create_mark_test_funcs}
\title{Given the names of mark test functions and the required mark statistics,
produce mark test functions as closures.}
\usage{
  create_mark_test_funcs(mtf_name, mark_stat_l)
}
\description{
  Given the names of mark test functions and the required
  mark statistics, produce mark test functions as closures.
}

