% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtf.r
\name{create_mtfs_and_weights}
\alias{create_mtfs_and_weights}
\title{Create mark test functions and weights}
\usage{
create_mtfs_and_weights(marks, mtf_name, edge_corr, one_per_lambda2)
}
\arguments{
\item{marks}{A vector of marks on which the mark test functions and their
scaling coefficients will be based.}

\item{mtf_name}{A vector containing the names of the mark test functions.}

\item{edge_corr}{A vector of edge correction coefficients for point
pairs.}

\item{one_per_lambda2}{The value of 1 / lambda^2 where lambda^2 is the
estimate of the intensity squared. A scalar.}
}
\value{
A list containing a list of mark tests functions (mtf_func_l) and
  a weight matrix (weight_m). Each mark test function takes two numeric
  vectors containing the marks of the points x_1 and the points x_2,
  respectively, from point pairs (x_1, x_2). The functions are in the
  same order as mtf_name. Each column of the weight matrix is reserved
  for use with the corresponding mark test function. The columns are in
  the same order as mtf_name. Each row corresponds to one pair of points.
}
\description{
Create a list of mark test functions and a matrix of weights.
}

