% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/radius.r
\name{create_r_max}
\alias{create_r_max}
\title{Create the maximum radius given the pattern.}
\usage{
create_r_max(min_window_side, recommended_r_max_ratio = 0.25)
}
\arguments{
\item{min_window_side}{The minimum window side length.}

\item{recommended_r_max_ratio}{The 'recommended' r_max / min_window_size ratio.}
}
\description{
Create the maximum radius given the pattern.
}

