% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/summ-func.r
\name{disc_param}
\alias{disc_param}
\title{Finds the radius and centre of a circle defined by a polygonial boundary}
\usage{
disc_param(x, ...)
}
\arguments{
\item{x}{An owin object.}

\item{...}{Ignored.}
}
\description{
Finds the radius and centre of a circle defined by a polygonial boundary
}

