\name{index_within_radius}
\alias{index_within_radius}
\title{Pick indices of separate point pairs that are within r_max of each other.}
\usage{
  index_within_radius(dist_m, r_max)
}
\value{
  A matrix of indices, row and column values separately.
}
\description{
  Pick indices of separate point pairs that are within
  r_max of each other.
}

