% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/summ-func.r
\name{is.disc}
\alias{is.disc}
\title{Checks if a window object is a circle}
\usage{
is.disc(x, digits = 6, ...)
}
\arguments{
\item{x}{An object to be checked.}

\item{digits}{The accuracy to check whether the points at the polygonial boundary
are on the arch of a circle.}

\item{...}{Ignored.}
}
\description{
Checks if a window object is a circle
}

