% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/summ-func.r
\name{K_f}
\alias{K_f}
\title{Calculates the cumulated K_f-estimates.}
\usage{
K_f(...)
}
\arguments{
\item{...}{Arguments to be passed to rho_f.}
}
\description{
Calculates the cumulated K_f-estimates.
}

