\name{mark_distr_stats}
\alias{mark_distr_stats}
\title{Adds only those mark statistics on the result list which are needed by
the K_f functions to be calculated from the given f.}
\usage{
  mark_distr_stats(marks, mtf_name)
}
\description{
  Adds only those mark statistics on the result list which
  are needed by the K_f functions to be calculated from the
  given f.
}

