% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/summ-func.r
\name{one_per_lambda_squared}
\alias{one_per_lambda_squared}
\title{Estimate 1 / lambda ^ 2.}
\usage{
one_per_lambda_squared(pattern, use_biased_lambda2)
}
\arguments{
\item{pattern}{A \code{\link[spatstat]{ppp}} object.}

\item{use_biased_lambda2}{A logical. Use the unbiased estimator or not.}
}
\description{
Notice that this estimator is biased.
}

