\name{pairs_within_max_radius}
\alias{pairs_within_max_radius}
\title{Pick indices of separate point pairs that are within r_max of each other.}
\usage{
  pairs_within_max_radius(dist_m, r_max)
}
\value{
  A matrix of indices, row and column values separately.
}
\description{
  Pick indices of separate point pairs that are within
  r_max of each other.
}

