\name{prepare_mark_test_funcs}
\alias{prepare_mark_test_funcs}
\title{Given the names of mark test functions and the required mark statistics,
produce mark test functions as closures.}
\usage{
  prepare_mark_test_funcs(mtf_name, mark_stat)
}
\description{
  Given the names of mark test functions and the required
  mark statistics, produce mark test functions as closures.
}

