\name{round_nonneg_into_int}
\alias{round_nonneg_into_int}
\title{Round non-negative numbers and convert them to integers.}
\usage{
  round_nonneg_into_int(x)
}
\arguments{
  \item{x}{A numeric type vector with only non-negative
  values.}
}
\value{
  An integer type vector with the given values rounded.
}
\description{
  Round non-negative numbers and convert them to integers.
}

