\name{summ_func}
\alias{summ_func}
\title{Summary functions for one pattern.}
\usage{
  summ_func(..., n_perm = 0L)
}
\value{
  A matrix with dimensions: summ_func, r.
}
\description{
  Uses the function
  \code{\link{summ_func_random_labelling}}. Eats the number
  of permutations the user has given and replaces it with
  zero.
}

