\name{disc_param}
\alias{disc_param}
\title{Finds the radius and centre of a circle defined by a polygonial boundary}
\usage{
  disc_param(x, ...)
}
\arguments{
  \item{x}{An owin object.}

  \item{...}{Ignored.}
}
\description{
  Finds the radius and centre of a circle defined by a
  polygonial boundary
}

