\name{is.disc}
\alias{is.disc}
\title{Checks if a window object is a circle}
\usage{
  is.disc(x, digits = 6, ...)
}
\arguments{
  \item{digit}{The accuracy to check whether the points at
  the polygonial boundary are on the arch of a circle.}
}
\description{
  Checks if a window object is a circle
}

