% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtf.r
\name{mark_distr_stats}
\alias{mark_distr_stats}
\title{Statistics of the mark distribution}
\usage{
mark_distr_stats(marks, mtf_name)
}
\arguments{
\item{marks}{A vector of marks of a point pattern.}

\item{mtf_name}{The name of mark test function.}
}
\description{
Adds only those mark statistics on the result list which are needed by
the K_f functions to be calculated from the given f.
}

