\name{min_window_side_length}
\alias{min_window_side_length}
\title{Length of the shorter window side.}
\usage{
  min_window_side_length(pattern)
}
\arguments{
  \item{pattern}{A \code{\link[spatstat]{ppp}} object that
  has a rectangular window.}
}
\value{
  A scalar value giving the length of the shorter edge of
  the window.
}
\description{
  Length of the shorter window side.
}

