% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/summ-func.r
\name{summ_func}
\alias{summ_func}
\title{Summary functions for one pattern.}
\usage{
summ_func(..., nsim = 0L)
}
\arguments{
\item{...}{Currently unused.}

\item{nsim}{Ignored.}
}
\value{
A matrix with dimensions: summ_func, r.
}
\description{
Uses the function \code{\link{summ_func_random_labelling}}. Eats the
number of permutations the user has given and replaces it with zero.
}
\seealso{
summ_func_random_labelling
}

